Revamp MAUI sample as a control gallery
Replace the single MainPage smoke test with a Shell flyout-based
gallery, one page per Compose-backed handler scenario:

  * Pages/CounterPage          - Button + Label counter (preserved
                                  from the dotnet new maui template).
  * Pages/ButtonsPage          - exercises every ButtonHandler mapper
                                  (Text / TextColor / Background /
                                  HorizontalLayoutAlignment), default
                                  vs themed colors, hug vs fill width.
  * Pages/LabelsPage           - exercises every LabelHandler mapper
                                  (Text / TextColor / Font / Horizontal
                                  TextAlignment / HorizontalLayoutAlignment).
  * Pages/EntriesPage          - exercises every EntryHandler mapper
                                  (Text echo, Placeholder, IsPassword,
                                  Keyboard - Numeric/Email/Url/Telephone,
                                  TextColor + Font, IsReadOnly,
                                  HorizontalLayoutAlignment).
  * Pages/ImageAspectsPage     - same dotnet_bot.png rendered with each
                                  Aspect (AspectFit / AspectFill / Fill /
                                  Center) so the cropping/letterbox/
                                  stretch differences are visible
                                  side-by-side.
  * Pages/ImageSourcesPage     - one row per ImageSource subtype
                                  (FileImageSource / UriImageSource /
                                  StreamImageSource / FontImageSource).

dotnet_bot.png is duplicated under Resources/Raw/ so the Stream
demo can open it via FileSystem.OpenAppPackageFileAsync. The
FontImageSource demo reuses the OpenSansSemibold font already
registered in MauiProgram.

AppShell switches from a single ShellContent to a FlyoutItem per
demo (FlyoutBehavior=Flyout); the old MainPage is removed.
