+++ src/Microsoft.AndroidX.Compose.Maui.Sample/AppShell.xaml
@@ -3,12 +3,39 @@
     x:Class="Microsoft.AndroidX.Compose.Maui.Sample.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:Microsoft.AndroidX.Compose.Maui.Sample"
-    Title="Microsoft.AndroidX.Compose.Maui.Sample">
+    xmlns:pages="clr-namespace:Microsoft.AndroidX.Compose.Maui.Sample.Pages"
+    Title="AndroidX.Compose backend"
+    FlyoutBehavior="Flyout">
 
-    <ShellContent
-        Title="Home"
-        ContentTemplate="{DataTemplate local:MainPage}"
-        Route="MainPage" />
+    <!--
+        Control gallery — one FlyoutItem per Compose-backed handler
+        scenario. The Shell chrome itself (toolbar, hamburger menu)
+        keeps stock MAUI rendering since `UseAndroidXCompose()` only
+        replaces leaf controls (Button, Entry, Image, Label).
+    -->
+
+    <FlyoutItem Title="Counter" Route="counter">
+        <ShellContent ContentTemplate="{DataTemplate pages:CounterPage}" />
+    </FlyoutItem>
+
+    <FlyoutItem Title="Buttons" Route="buttons">
+        <ShellContent ContentTemplate="{DataTemplate pages:ButtonsPage}" />
+    </FlyoutItem>
+
+    <FlyoutItem Title="Labels" Route="labels">
+        <ShellContent ContentTemplate="{DataTemplate pages:LabelsPage}" />
+    </FlyoutItem>
+
+    <FlyoutItem Title="Entries" Route="entries">
+        <ShellContent ContentTemplate="{DataTemplate pages:EntriesPage}" />
+    </FlyoutItem>
+
+    <FlyoutItem Title="Image: Aspects" Route="image-aspects">
+        <ShellContent ContentTemplate="{DataTemplate pages:ImageAspectsPage}" />
+    </FlyoutItem>
+
+    <FlyoutItem Title="Image: Sources" Route="image-sources">
+        <ShellContent ContentTemplate="{DataTemplate pages:ImageSourcesPage}" />
+    </FlyoutItem>
 
 </Shell>
