--- src/Microsoft.AndroidX.Compose.Maui.Sample/MainPage.xaml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Microsoft.AndroidX.Compose.Maui.Sample.MainPage"
-             Title="AndroidX.Compose backend">
-
-    <ScrollView>
-        <VerticalStackLayout
-            Padding="30,0"
-            Spacing="25">
-            <!-- FileImageSource: takes the IFileImageSource fast path
-                 (resolves to R.drawable.dotnet_bot and renders via
-                 Compose's painterResource(int)). -->
-            <Image
-                Source="dotnet_bot.png"
-                HeightRequest="185"
-                Aspect="AspectFit"
-                SemanticProperties.Description="dot net bot in a submarine number ten" />
-
-            <!-- UriImageSource: exercises the general MAUI pipeline.
-                 UriImageSourceService downloads on a worker thread,
-                 the handler wraps the resulting Drawable as a Compose
-                 BitmapPainter. Uses the .NET brand logo (small,
-                 stable, CDN-cached). -->
-            <Image
-                Source="https://raw.githubusercontent.com/dotnet/brand/main/logo/dotnet-logo.png"
-                HeightRequest="120"
-                Aspect="AspectFit"
-                SemanticProperties.Description="dot net logo loaded over HTTPS" />
-
-            <Label
-                Text="Hello, World!"
-                Style="{StaticResource Headline}"
-                SemanticProperties.HeadingLevel="Level1" />
-
-            <Label
-                Text="Welcome to &#10;.NET MAUI on Jetpack Compose"
-                Style="{StaticResource SubHeadline}"
-                SemanticProperties.HeadingLevel="Level2"
-                SemanticProperties.Description="Welcome to dot net MAUI on Jetpack Compose" />
-
-            <Entry
-                x:Name="NameEntry"
-                Placeholder="Type your name"
-                TextChanged="OnNameTextChanged"
-                HorizontalOptions="Fill" />
-
-            <Label
-                x:Name="GreetingLabel"
-                Text="Greeting will appear here."
-                HorizontalOptions="Fill" />
-
-            <Entry
-                x:Name="SecretEntry"
-                Placeholder="Password (masked)"
-                IsPassword="True"
-                HorizontalOptions="Fill" />
-
-            <Entry
-                x:Name="DigitsEntry"
-                Placeholder="Numeric IME"
-                Keyboard="Numeric"
-                HorizontalOptions="Fill" />
-
-            <!-- Exercises MapTextColor + MapFont (size + bold) on the
-                 input — analogue of the gallery's TextStyleOverrideDemo. -->
-            <Entry
-                x:Name="StyledEntry"
-                Text="Big bold blue"
-                TextColor="#2196F3"
-                FontSize="22"
-                FontAttributes="Bold"
-                HorizontalOptions="Fill" />
-
-            <!-- Exercises MapIsReadOnly — focus + selection still work,
-                 keyboard suppressed. -->
-            <Entry
-                x:Name="ReadOnlyEntry"
-                Text="Read-only (no keyboard)"
-                IsReadOnly="True"
-                HorizontalOptions="Fill" />
-
-            <Button
-                x:Name="CounterBtn"
-                Text="Click me"
-                BackgroundColor="{StaticResource Primary}"
-                TextColor="White"
-                SemanticProperties.Hint="Counts the number of times you click"
-                Clicked="OnCounterClicked"
-                HorizontalOptions="Fill" />
-        </VerticalStackLayout>
-    </ScrollView>
-
-</ContentPage>
