+++ src/Microsoft.AndroidX.Compose.Maui.Sample/Pages/LabelsPage.xaml.cs
@@ -1,0 +1,16 @@
+namespace Microsoft.AndroidX.Compose.Maui.Sample.Pages;
+
+/// <summary>
+/// Labels demo — exercises every mapper on
+/// <see cref="Microsoft.AndroidX.Compose.Maui.Handlers.LabelHandler"/>:
+/// Text, TextColor, Font (size + bold), HorizontalTextAlignment, and
+/// HorizontalLayoutAlignment.
+/// </summary>
+public partial class LabelsPage : ContentPage
+{
+    /// <summary>Build the page.</summary>
+    public LabelsPage()
+    {
+        InitializeComponent();
+    }
+}
